# packages/contracts/gen/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "platform aggregate API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/rbac/permissions": {
      "post": {
        "x-request-body-limit": 4096,
        "operationId": "createPermission",
        "summary": "add a unique permission to the catalog",
        "x-required-permission": "role:update:any",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 120,
                    "pattern": "^[a-z0-9_]+:[a-z0-9_]+:[a-z0-9_]+$",
                    "example": "report:export:any"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "created (or already exists)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          },
          "400": {
            "description": "invalid name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          },
          "409": {
            "description": "duplicate permission name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listPermissions",
        "summary": "the compile-time catalog persisted in the db",
        "x-required-permission": "permission:read:any",
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/EnvelopeOK"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "items": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Permission"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rbac/permissions/{name}": {
      "delete": {
        "operationId": "deletePermission",
        "summary": "delete an unused permission from the catalog",
        "x-required-permission": "role:update:any",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 120
            }
          }
        ],
        "responses": {
          "200": {
            "description": "deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          },
          "404": {
            "description": "unknown permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          },
          "409": {
            "description": "permission is in use",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rbac/roles": {
      "post": {
        "x-request-body-limit": 32768,
        "operationId": "createRole",
        "x-required-permission": "role:create:any",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/EnvelopeOK"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Role"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "duplicate role name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listRoles",
        "x-required-permission": "role:read:any",
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/EnvelopeOK"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "items": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Role"
                              }
                            },
                            "meta": {
                              "$ref": "#/components/schemas/EnvelopeMeta"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rbac/roles/{id}": {
      "patch": {
        "x-request-body-limit": 32768,
        "operationId": "updateRole",
        "summary": "rename/describe and/or sync the permission set; bumps affected users' ver",
        "x-required-permission": "role:update:any",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/RoleInput"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "permissions": {
                        "type": "array",
                        "maxItems": 200,
                        "uniqueItems": true,
                        "description": "full replacement set; unknown permission -> 400",
                        "items": {
                          "type": "string",
                          "maxLength": 120
                        }
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updated",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/EnvelopeOK"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Role"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "unknown permission in set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          },
          "404": {
            "description": "unknown role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteRole",
        "x-required-permission": "role:delete:any",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "fallbackRoleId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          },
          "404": {
            "description": "unknown role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rbac/users/{id}/roles": {
      "get": {
        "operationId": "getUserRoles",
        "summary": "list roles assigned to a user",
        "x-required-permission": "role:read:any",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          }
        }
      },
      "put": {
        "x-request-body-limit": 16384,
        "operationId": "setUserRoles",
        "summary": "replace the role set assigned to a user (bumps their ver)",
        "x-required-permission": "role:update:any",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "roleIds"
                ],
                "properties": {
                  "roleIds": {
                    "type": "array",
                    "maxItems": 100,
                    "uniqueItems": true,
                    "items": {
                      "type": "integer",
                      "format": "int64"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          },
          "400": {
            "description": "unknown role in set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rbac/internal/claims/{sub}": {
      "get": {
        "operationId": "resolveClaims",
        "summary": "internal API — perms[] + ver for a subject; requires the shared internal secret",
        "x-internal": true,
        "security": [],
        "parameters": [
          {
            "name": "sub",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/EnvelopeOK"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "perms": {
                              "type": "array",
                              "items": {
                                "type": "string",
                                "maxLength": 120
                              }
                            },
                            "ver": {
                              "type": "integer",
                              "format": "int64"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "missing/wrong internal secret",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "EnvelopeMeta": {
        "type": "object",
        "required": [
          "limit",
          "offset",
          "total"
        ],
        "properties": {
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "EnvelopeOK": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "maxLength": 255
          },
          "data": {
            "description": "endpoint-specific payload"
          }
        }
      },
      "EnvelopeFail": {
        "type": "object",
        "required": [
          "success",
          "message",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "maxLength": 255
          },
          "error": {
            "type": "string",
            "maxLength": 120
          }
        }
      },
...
      "RoleInput": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 60
          },
          "description": {
            "type": "string",
            "maxLength": 300
          },
          "color": {
            "type": "string",
            "maxLength": 7,
            "pattern": "^#[0-9a-fA-F]{6}$",
            "default": "#6366f1"
          },
          "icon": {
            "type": "string",
            "maxLength": 60,
            "default": "shield"
          },
          "archived": {
            "type": "boolean",
            "default": false
          },
          "permissions": {
            "type": "array",
            "maxItems": 200,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "maxLength": 120
            }
          }
        }
      },
      "Permission": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "roleCount": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "maxLength": 60
          },
          "description": {
            "type": "string",
            "maxLength": 300
          },
          "color": {
            "type": "string",
            "maxLength": 7
          },
          "icon": {
            "type": "string",
            "maxLength": 60
          },
          "archived": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "userCount": {
            "type": "integer",
            "format": "int64"
          },
          "system": {
            "type": "boolean"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 120
            }
          }
        }
      },
...
    }
  }
}